refactor(test): MeshWeaver.Content.Test drops its MeshWeaver.AI project reference
Four files coupled core's content suite to the AI assembly, each for a different reason,
so each needed a different answer:

- AgentFileParserTest (613 lines) and ThreadMessageToolTest are pure AI tests that had
  simply landed in the wrong project. Moved to MeshWeaver.AI.Test verbatim.

- ContributedParserPriorityTest is a test of FileFormatParserRegistry's ORDERING contract
  that happened to use AgentFileParser as its concrete contributor. Pinning a registry
  against one particular module's parser both coupled core to that module and quietly
  narrowed the assertion to "this one client works". The core half now uses a test-local
  ProbeFileParser, so it holds for every contributor including ones that do not exist yet;
  the real-parser proof moves to MeshWeaver.AI.Test.AgentParserContributionTest, which
  asserts the thing only the AI module can — that an agent file survives as
  AgentConfiguration rather than degrading into a hollow Markdown node.

- SpacesInFileNameTest spans both: its five MarkdownExtractor_* cases test
  MarkdownReferenceExtractor (MeshWeaver.AI), while the UcrPrefixResolver and
  ContentAutocompleteProvider cases are core. Split accordingly; the extracted half needs
  no mesh (every method under test is static and pure) so it is a plain xUnit class.

🚨 One using of MeshWeaver.AI.Completion REMAINS in Content.Test, deliberately, with a
comment saying why: FuzzyScorer ships in MeshWeaver.Data but declares a MeshWeaver.AI
namespace. That mismatch is filed as #2344 — the using is not an AI dependency and removing
it does not remove one. It cost a build cycle here precisely because it looks like one.

Verified: both projects build -c Release -warnaserror with 0 Warning(s)/0 Error(s), and both
sides run green (18 Content + 46 AI).

No What's New entry: test-only refactor, no user-visible effect.

Refs #2276
