+++ test/MeshWeaver.AI.Test/AgentFileParserTest.cs
@@ -1,11 +1,10 @@
 using System.Threading.Tasks;
-using MeshWeaver.AI;
 using MeshWeaver.AI.Persistence;
 using MeshWeaver.Hosting.Persistence.Parsers;
 using MeshWeaver.Mesh;
 using Xunit;
 
-namespace MeshWeaver.Content.Test;
+namespace MeshWeaver.AI.Test;
 
 /// <summary>
 /// Unit tests for AgentFileParser - parsing and serialization of agent markdown files with YAML front matter.
